+++ WireAnalytics/Sources/WireAnalytics/Events/AnalyticsEvent+Predefined.swift
@@ -110,46 +110,3 @@ public extension AnalyticsEvent {
         .init(name: "calling.call_quality_review", segmentation: review.segmentation)
     }
 }
-
-/// User review for call quality.
-
-public enum CallQualitySurveyReview {
-
-    /// The survey was not displayed.
-    case notDisplayed(reason: IgnoreReason, duration: TimeInterval)
-
-    /// The survey was answered by the user.
-    case answered(score: Int, duration: TimeInterval)
-
-    /// The survey was dismissed.
-    case dismissed(duration: TimeInterval)
-
-    public enum IgnoreReason: String {
-        case callTooShort = "call-too-short"
-        case muted
-    }
-
-    var segmentation: Set<SegmentationEntry> {
-        switch self {
-        case .notDisplayed(let reason, let duration):
-            [
-                .callLabel("not-displayed"),
-                .callDuration(duration),
-                .callIgnoreReason(reason.rawValue)
-            ]
-
-        case .answered(let score, let duration):
-            [
-                .callLabel("answered"),
-                .callScore(score),
-                .callDuration(duration)
-            ]
-
-        case .dismissed(let duration):
-            [
-                .callLabel("dismissed"),
-                .callDuration(duration)
-            ]
-        }
-    }
-}
